some cleanup on update features
