+++ resources/settings.xml
@@ -32,7 +32,7 @@
         <setting label="$ADDON[plugin.video.themoviedb.helper 32086]" type="bool" id="library_autoupdate" default="false" />
         <setting label="$ADDON[plugin.video.themoviedb.helper 32130]" subsetting="true" type="enum" values="$HOURS" id="library_autoupdate_hour" enable="eq(-1,True)" />
         <setting label="$ADDON[plugin.video.themoviedb.helper 32131]" subsetting="true" type="action" action="RunScript(plugin.video.themoviedb.helper,monitor_userlist)" id="monitor_userlist"  enable="eq(-2,True)"/>
-        <setting label="$ADDON[plugin.video.themoviedb.helper 32154]" subsetting="true" type="action" action="RunScript(plugin.video.themoviedb.helper,library_autoupdate,busy_dialog)" enable="eq(-3,True)"/>
+        <setting label="$ADDON[plugin.video.themoviedb.helper 32154]" subsetting="true" type="action" action="RunScript(plugin.video.themoviedb.helper,library_autoupdate,busy_dialog,force)" enable="eq(-3,True)"/>
         <setting label="$LOCALIZE[10039]" type="lsep"/>
         <setting label="$ADDON[plugin.video.themoviedb.helper 32077]" type="bool" id="ignore_folderchecking" default="false" />
     </category>
